seo-ops/integrations/requirements.txt
google-auth>=2.20,<3
google-auth-oauthlib>=1.0,<2
google-api-python-client>=2.90,<3
google-analytics-data>=0.18,<1
